refactor(kilo,opencode): split models.ts into model modules
Both backends share a forked provider-bucket catalog, so they get a
parallel models/ layout:
  - types        — raw + public catalog shapes
  - catalog      — TTL cache, raw→catalog parsing, fetch entry point
  - resolve      — query parse/match, provider-id guess, model resolver
  - presentation — quick-pick, settings buttons, summary/list, errors

kilo/models.ts was 804 lines, opencode/models.ts was 784. The two stay
~95% identical; the documented divergences are preserved exactly
(opencode: isFreeModel OR-semantics, no resolve fast-path, 48-char
callback-safe IDs, 4-pick budget, "OpenCode" branding). index.ts in each
re-exports the same public surface; importers + tests updated to the
barrel. Verified: kilo + opencode model suites pass.
